changed tabs to spaces

+++ plugin-init.py
@@ -7,11 +7,11 @@ new_name = sys.argv[1]
 
 
 def update_folder(path, current_name, new_name):
-	os.rename(path + current_name, path + new_name)
-	if( os.path.exists(path + new_name)):
-		print ("OK: " + path + new_name + " updated")
-	else:
-		print ("Error: " +  + path + new_name + "failed to update")
+    os.rename(path + current_name, path + new_name)
+    if( os.path.exists(path + new_name)):
+        print ("OK: " + path + new_name + " updated")
+    else:
+        print ("Error: " +  + path + new_name + "failed to update")
 
 
 print ("updating plugin with new name: " + new_name + "\n")
@@ -27,24 +27,25 @@ update_folder("pkg/files-win/plugins/scripts/", current_name, new_name)
 print "Creating and updating " + new_name + ".xml with new name & paths"
 old_xml_path = "pkg/" + current_name + ".xml"
 new_xml_path = "pkg/" + new_name + ".xml"
-old_xml = open(old_xml_path, 'r')
-new_xml = open(new_xml_path, 'w')
+if (os.path.exists(old_xml_path)):
+    old_xml = open(old_xml_path, 'r')
+    new_xml = open(new_xml_path, 'w')
 
-for line in old_xml:
-    new_xml.write(line.replace(current_name, new_name))
+    for line in old_xml:
+        new_xml.write(line.replace(current_name, new_name))
 
-old_xml.close()
-new_xml.close()
+    old_xml.close()
+    new_xml.close()
 
 if (os.path.exists(new_xml_path)):
-	print (new_xml_path + " exists")
-	if (os.remove(old_xml_path)):
-		print ("Removed old xml file: " + old_xml_path)
-	else:
-		print ("Unable to remove old xml: " + old_xml_path)
-		print ("Try deleting file manually")
+    print (new_xml_path + " exists")
+    if (os.remove(old_xml_path)):
+        print ("Removed old xml file: " + old_xml_path)
+    else:
+        print ("Unable to remove old xml: " + old_xml_path)
+        print ("Try deleting file manually")
 else:
-	print (new_xml_path +"does not exist?")
+    print (new_xml_path +"does not exist?")
